Add TaskScheduler
Designing a scheduler and dispatcher for a graph of tasks. The goal is to respect dependencies while having a performant concurrent execution when possible.
The two classes are TaskScheduler and TaskProxy. TaskScheduler is a high level convenience API to prepare and kick off the execution. TaskProxy is a node in the task graph to put the actual task block in and connect to other nodes to construct the dependency graph.
Construction of the dependency graph is actually the difficult part which is not addressed in this change. TaskScheduler and TaskProxy is just the groundwork.