Medusa algorithm exploration
This is a medusa algorithm exploration I did a year ago to connect inputs and ouptuts of processes.
Medua (or a jellyfish) is a free-floating, unconnected being with tentacles. For this excercise we have a special breed which has tentacles categorized as inputs and outputs. Inputs can be anchored to some static structure like an island, outputs can only be connected to other medusas' inputs or remain unconnected. These medusas hate to dance in circles so the structure they create by connecting tentacles is acyclic (i.e. a DAG).
Why do we care?  Some systems are putting a burden of graph construction on user/integrtator. Be it a build graph or other hierarical execution graph. It is a win if, instead, the user can describe inputs and outputs of each process independently from other processes and let the machine find the connections between the nodes.
The next step would be to adapt the algorithm to the needs of "replay" tool and construct the graph for TaskScheduler to execute concurrently with GCD