Medusa 3 excercise
Medusa 1 & 2 put restrictions on static inputs (not connected to other medusa's outputs) to be easily distinguishable from outputs. This is an optimization which has its root in build system where sources and output build are separated. With medusa 3, that optimization is removed, requiring a hard check for unonnected inputs, which adds an expense but removes the requirement to separate input from output location.