Add Obj-C version of medusa
One step closer to integrate with replay and scheduler is to move the algorithm into Obj-C domain. There seems to be some performance degradation compared to STL C++ version which should be investigated and mitigated.
Fixed one inefficency in algorithm in all versions to break early in look-ahead consumers when the first not satisifed input is detected - stop iterating because next input status will not change the outcome: the node cannot be executed in next iteration. This cuts execution time by ~20% in C++ version but has much bigger impact on Obj-C version: more than halves the time spent  in execute_medusa_list(), e.g. 40s to 16s.