Complete graph construction in Objective-C, aka medusa algorithm
Completed two versions of medusa algorithm in Objective C:
- single-threaded recursive execution
- concurrent execution with TaskScheduler
Concurrent implementation should be mostly ready for integration with replay tool
Recursive implementation is less interesting for replay but might be super useful for generating and serializing a graph for an external scheduler/executor engine.

The next steps will be cleanup and integration with replay tool.