FileTree - convert list of paths into directory tree
[Resubmitting from the coreect private account. This project is developed outside of work hours]
Code experiements for converting a flat list of paths into a tree. Final piece for building dependencies for "replay" tool. Each node in a tree represents a file or directory and may have an associated producer in the task graph. Double linked list will allow us to find all producers from given leaf to the root and therefore task dependencies.