Integrate FileTree into medusa
Replace string-based PathSpec with the FileNode in the file tree in medusa code.
Each input and output path string needs to be converted to the FileNode by calling FindOrInsertFileNodeForPath(). There is only one shared tree where both inputs and outputs nodes are and the same paths are deduplicated. By specifying a producer in the output node we also specify it in the input node so medusa algorithm can build a dependency graph between tasks.
The file tree is supposed to be kept alive for the life of the process and tasks keep weak references to these nodes.