Major update to concurrent execution
This change brings serveral weekend's worth of code to "replay". The  new engine to analyze dependenices and schedule tasks is merged into to the main tool. It allows describing hundreds or thousands of tasks in the playlist and the "medusa" algorithim will construct an execution graph based on input and output paths as long as the reasonable rules are followed. The main advantage is that you don't need to describe which task must follow which or what tasks may be performed concurrently. Replay will analyze the consumed and produced files and schedule tasks accrodingly to ensure producers will be executed before relevant consumers.