Add action streaming
Add support for streaming actions via stdin pipe. The format is different and execution is concurrent by default but without dependency analysis.
Changed "raw content" option to "raw" for file creation - the desire is to have a single option word without space for streaming format and keep it the same as in playlist format.
Added "stdout" boolean flag for "execute" action to allow supressing the output from a child tool.