Add @autoreleasepools
Stress testing revealed a problem that unreleased NSPipe objects were holding on too many file descriptiors and finally started failing to execute NSTasks. This is becuase the GCD queue built-in autoreleasing was not frequent enough. Add individual @autoreleasepools around each dispatched task as well as the whole body of HandleActionStep() function, which may be called millions of times if you stream tasks.