Add verification for concurrent execution with dependencies
Added detection of dependency cycles in action graph for concurrent execution. This revealed a couple of bugs, which were fixed.

The completness verification happens after execution. The algorithm does not allow doing it upfront because the execution mutates the graph. It would require cloning the graph to run verification. It is expensive but might be a good idea with explicit request from client - for long executing graphs and to prevent running anything if there is a cycle detected.

Added more tests for better code coverage.