Add "dispatch" tool helper
"dispatch" tool is designed to simplify scheduling ad hoc tasks for execution in "replay" without the need to create a playlist. "replay" gains the ability to run in a server mode and receive messages from the "dispatch" driver.

If you are familiar with programming on Apple platforms you can think of this setup as "replay" server acting as a GCD service and "dispatch" tool being the equivalent of dispatch_async API on concurrent queue. The difference is that instead of dispatching blocks in native code you are dispatching child binary invocations for concurrent execution.

If you have a script with several tasks executed serially but they are independent and be parallelized you can take advantage of this tool.
Suppose you execute two long-ish tasks serially in your script:
 /path/to/tool1 param1 param2
 /path/to/tool2 param3

You can replace it with:
dispatch "job-id" execute /path/to/tool1 param1 param2
dispatch "job-id" execute /path/to/tool2 param3
dispatch "job-id" wait
tool1 and tool2 will be executed concurrently by "replay" server in the background and your sctipr will wait for both of them to finish before continuing the execution.