Add stdout and stderr options
Add options to redirect stdout and stderr to chosen files. The primary reason for adding this feature is redirecting the output when "replay" is running in server mode and accepts "dispatch" requests. It can also be useful for keeping the logs in expected locations when the exeuction is unattended. stderr output is always written in the order the tasks are executed. There is no option for concurrent execution to keep the task errors in the order of schedulting like we have for stdout with "--ordered-output" option.