A couple of fixes for dispatch
When installed in default /usr/local/bin/ location, and invoked without full path to the tool, "dispatch" could not find "replay". Obtain real executable location in the code and not rely at 0th argument.
"wait" action was incorrectly printing that server was not running for given batch when in fact it was.