Add support for specifying input file list with --inputs param
Most common type of file list with one path per line is Xcode's xcfilelist. It may contain environment variables in form of ${VAR} or Xcode-specific $(VAR).
For fingerprint `--inputs` it does not have to be an xcfilelist, it can be any plain text file with one path per line, optionally with env variables, e.g.:
```
${HOME}/foo.txt
/Users/${USER}/bar.txt
```