xattr readonly workaround
Setting or clearing xattr respects file's read-only permission.
The end result is that we cannot set fingerprint extended attribute cache for read-only files.

Now, what does it mean to modify the file's extended attributes? Does it touch the file? No. It stores or removes associated data in file system, which is not even with the file. Extended attribute mutation does not change anything about the file: not data, not file inode, not modification date. So we may make an argument that file readonly flag should not really apply to xattrs. This is what this change is about: we make the readonly file writable temporarily, change the xattr and then restore original permissions. Nothing changes about the file except that now we have a fingerprint hash in the extended attributes. A cached hash for read-only file is even more valuable than some volatile writable file: the read-only file is not supposed to change so calculating hash for it each time is wasteful and caching it is highly beneficial.